requirements.txt
biopython
joblib
networkx
numpy
matplotlib
pandas
python-edlib
seaborn
scikit-learn==0.21.3
setuptools
